+++ Changelog.md
@@ -7,6 +7,7 @@ OpenCore Changelog
 - Fixed FS discovery on NVMe with legacy drivers
 - Added `DirectGopCacheMode` option for FB cache policy
 - Added `KeyFiltering` option to workaround buggy KB drivers
+- Added tool and custom entry separation in audio assistant
 
 #### v0.5.6
 - Various improvements to builtin text renderer